feat: (6/n) split `relay-helper.ts` to multiple files
