+++ components/quiz/EditClozeQuestion.tsx
@@ -2,7 +2,7 @@ import { EditClozeQuestionFragment$key } from "@/__generated__/EditClozeQuestion
 import { EditClozeQuestionMutation } from "@/__generated__/EditClozeQuestionMutation.graphql";
 import { lecturerAllSkillsQuery } from "@/__generated__/lecturerAllSkillsQuery.graphql";
 import { MediaRecordSelector$key } from "@/__generated__/MediaRecordSelector.graphql";
-import { questionUpdaterClosure } from "@/src/relay-helpers";
+import { questionUpdaterClosure } from "@/src/relay-helpers/question";
 import { useParams } from "next/navigation";
 import { useCallback, useState } from "react";
 import { graphql, PreloadedQuery, useFragment, useMutation } from "react-relay";
